refactor: streamline WorkoutPage and introduce CreateYourOwnWorkoutPage
- Removed unused state and caching logic from WorkoutPage for improved clarity.
- Added CreateYourOwnWorkoutPage component to allow users to create custom workouts.
- Enhanced layout and filtering functionality across both pages for better user experience.
- Updated UI elements and structure for consistency and readability.
